@@ -4,8 +4,7 @@ import { PageModule } from './../shared/page.module';
 import { AccountComponent } from './account.component';
 import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
-import { RouterModule } from '@angular/router';
-import { HttpModule } from '@angular/http';
+import { HttpClientModule } from '@angular/common/http';
 import { AccountRoutingModule } from './account.routing.module';
 
 @NgModule({
@@ -13,7 +12,7 @@ import { AccountRoutingModule } from './account.routing.module';
         CommonModule,
         PageModule,
         FormsModule,
-        HttpModule,
+        HttpClientModule,
         AccountRoutingModule
     ],
     declarations: [
