@@ -5,8 +5,7 @@ import { CommonModule } from '@angular/common';
 import { ProfileComponent } from './profile.component';
 import { ProfileDescriptionComponent } from './profile-description.component';
 import { Ng2PageScrollModule } from 'ng2-page-scroll/ng2-page-scroll';
-// import { Angulartics2On } from 'angulartics2';
-import { HttpModule } from '@angular/http';
+import { HttpClientModule } from '@angular/common/http';
 import { HomeRoutingModule } from './home.routing.module';
 import { ProfileService } from './profile.service';
 import { PipesModule } from '../../pipes/pipes.module';
@@ -17,7 +16,7 @@ import { PipesModule } from '../../pipes/pipes.module';
         CommonModule,
         PageModule,
         Ng2PageScrollModule.forRoot(),
-        HttpModule,
+        HttpClientModule,
         HomeRoutingModule,
         PipesModule
     ],
