@@ -6,7 +6,7 @@ import { LoaderComponent } from './loader.component';
 import { PageModule } from './page.module';
 import { CommonModule } from '@angular/common';
 import { RouterModule } from '@angular/router';
-import { HttpModule } from '@angular/http';
+import { HttpClientModule } from '@angular/common/http';
 import { PasswordHeaderComponent } from './password-header.component';
 
 
@@ -14,7 +14,7 @@ import { PasswordHeaderComponent } from './password-header.component';
     imports: [
         CommonModule,
         RouterModule,
-        HttpModule,
+        HttpClientModule,
         PageModule
     ],
     exports: [
