@@ -1,15 +1,12 @@
 import { Injectable } from '@angular/core';
-import { Http, Headers, RequestOptions, Response } from '@angular/http';
-import { Observable } from 'rxjs/Rx';
-// import {AuthService} from './auth';
-import { links } from '../constants/config';
-// import {CommonService} from './common';
+import { HttpClient, HttpHeaders, HttpResponse } from '@angular/common/http';
+import { Observable } from 'rxjs';
 import { AuthService } from './auth.service';
 import { CommonService } from './common.service';
 
 @Injectable()
 export class LogService {
-    constructor(private auth: AuthService, private http: Http, private common: CommonService) {
+    constructor(private auth: AuthService, private http: HttpClient, private common: CommonService) {
 
     }
 
@@ -21,15 +18,17 @@ export class LogService {
                 Reason: reason
             });
             let url: string = this.common.getLogServer();
-            let headers: Headers = new Headers();
+            const headers = new HttpHeaders();
             headers.append('Authorization', this.auth.authheader);
             headers.append('Accept', 'application/json');
             headers.append('Content-Type', 'application/json');
-            let options: RequestOptions = new RequestOptions();
-            options.headers = headers;
+            let options = {
+                headers: headers,
+                observe: 'response' as const
 
+            }
             this.http.post(url, body, options)
-                .map(res => <any>res.json())
+                .map(res => <any>res)
                 .subscribe(
                 response => console.log(response),
                 error => this.handleError,
@@ -41,8 +40,8 @@ export class LogService {
 
     }
 
-    handleError(error: Response): any {
-        return Observable.throw(error.json().error || 'Server error');
+    handleError(error: HttpResponse<any>): any {
+        return Observable.throw(error.body.error || 'Server error');
     }
 
 }
