@@ -1,26 +1,13 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { Title } from '@angular/platform-browser';
-// import {AuthService} from '../../services/auth';
-// import {CommonService} from '../../services/common';
-// import * as _ from 'lodash';
-// import {PageHeader} from '../shared/page-header';
-// import {PageFooter} from '../shared/page-footer';
-// import {HomeService} from '../../services/home-service';
 import { links } from '../../constants/config';
 import { AuthService } from './../../services/auth.service';
 import { CommonService } from './../../services/common.service';
-import { Subscription, Observable } from 'rxjs/Rx';
-import { Http, Response, Headers, RequestOptions } from '@angular/http';
-
-// import { PageHeaderComponent } from './../shared/page-header.component';
-// import { PageFooterComponent } from './../shared/page-footer.component';
-
+import { HttpClient, HttpHeaders } from '@angular/common/http';
 @Component({
     selector: 'reports',
-    // providers: [AuthService, CommonService],
-    templateUrl: './reports.component.html'//,
-    // directives: [, PageHeaderComponent, PageFooterComponent]
+    templateUrl: './reports.component.html'
 })
 
 export class ReportsComponent implements OnInit {
@@ -40,7 +27,7 @@ export class ReportsComponent implements OnInit {
     examityEncryptedUserId: string;
     examityServer:string;
     examityLoginURL:string;
-    constructor(private http: Http, public auth: AuthService, public router: Router, public common: CommonService, public titleService: Title) {
+    constructor(private http: HttpClient, public auth: AuthService, public router: Router, public common: CommonService, public titleService: Title) {
         this.apiServer = this.common.getApiServer();
         this.nursingITServer = this.common.getNursingITServer();
     }
@@ -91,24 +78,26 @@ export class ReportsComponent implements OnInit {
 
     onClickExamityProfile(ssologin, encryptedUsername_val): void {
         let facultyAPIUrl = this.resolveFacultyURL(`${this.common.apiServer}${links.api.baseurl}${links.api.admin.examityProfileapi}`);
-        let examityObservable: Observable<Response> = this.setFacultyProfileInExamity(facultyAPIUrl);
+        let examityObservable  = this.setFacultyProfileInExamity(facultyAPIUrl);
         examityObservable.subscribe(response => {
-            this.examityEncryptedUserId = response.json();
+            this.examityEncryptedUserId = response.toString();
             encryptedUsername_val.value = this.examityEncryptedUserId
             ssologin.submit();
         }, error => console.log(error));
     }
 
-    setFacultyProfileInExamity(url: string): Observable<Response> {
+    setFacultyProfileInExamity(url: string)  {
         let self = this;
-        let options: RequestOptions = new RequestOptions();
-        let headers: Headers = new Headers({
+        const headers = new HttpHeaders({
             'Accept': 'application/json',
             'Content-Type': 'application/json',
             'Authorization': self.auth.authheader
         });
-        options.headers = headers;
-        options.body = '';
+        let options = {
+            headers : headers,
+            // body : ''
+            observe: 'response' as const
+        };
         return this.http.get(url, options);
     }
 
