@@ -1,5 +1,4 @@
-import { Component, Input, OnInit, Output, EventEmitter } from '@angular/core';
-import { NgIf } from '@angular/common';
+import { Component, OnInit, Output, EventEmitter } from '@angular/core';
 import { AuthService } from './../../services/auth.service';
 
 @Component({
@@ -29,7 +28,7 @@ export class ManageTestsMultiCampusComponent implements OnInit {
     }
 
     refreshSelectpicker(): void {
-        setTimeout(json => {
+        setTimeout(() => {
             if (/Android|webOS|iPhone|iPad|iPod|BlackBerry/i.test(navigator.userAgent))
                 $('.selectpicker').selectpicker('mobile');
             else
