@@ -1,16 +1,10 @@
 import { Injectable, Injector, ErrorHandler} from '@angular/core';
-import {Http} from '@angular/http';
 import {Router} from '@angular/router';
-// import {LogService} from '../services/log.service.service';
-// import {AuthService} from '../services/auth';
-// import {CommonService} from '../services/common';
-import { AuthService } from './../services/auth.service';
-import { CommonService } from './../services/common.service';
 import { LogService } from './../services/log.service';
 
 @Injectable()
 export class MyExceptionHandler implements ErrorHandler {
-    constructor(public injector: Injector, private log:LogService) {
+    constructor(public injector: Injector) {
         // console.log(this.router);
     }
 
